move the adapter service of the my-campaigns module to its own subdirectory (GUI)
