+++ artushima-web/src/app/my-campaigns/components/start-campaign/start-campaign.component.ts
@@ -1,17 +1,11 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, Input } from '@angular/core';
 import { Router } from '@angular/router';
-
-// Services
+import { MessageLevel } from 'src/app/core/model/message-level';
+import { RequestStatus } from 'src/app/core/model/request-status';
 import { MessagesService } from 'src/app/core/services/messages.service';
-import { MyCampaignsAdapterService } from '../../services/my-campaigns-adapter.service';
-
-// Constants
 import { DEFAULT_CAMPAIGN_START_DATE } from '../../constants/my-campaigns.constants';
 import { MSG_CAMPAIGN_CREATED } from '../../constants/my-campaings.messages';
-
-// Model
-import { RequestStatus } from 'src/app/core/model/request-status';
-import { MessageLevel } from 'src/app/core/model/message-level';
+import { MyCampaignsAdapterService } from '../../services/my-campaigns-adapter.service/my-campaigns-adapter.service';
 
 /**
  * The component displaying a form for starting a new campaign with
